Set basic widths
A single-column mobile layout is set by default.
If possible, it is replaced by a 960px 12-column grid.

The basic width is now fixed to 960px. However, all widths are relative
to a base width, and the grid can be made fluid by modifying this
base width (#page-wrapper {width: 100%;} by example).